Rename importer script to env.
