+++ test/unit/tests/collision.js
@@ -1,7 +1,7 @@
 var expect = require('expect.js');
 var sinon = require('sinon');
 
-var env = require('../../importer.js');
+var env = require('../../env.js');
 
 var Collision = env.Engine.Collision;
 var Obj = env.Engine.Object;
