+++ test/unit/tests/object.js
@@ -1,7 +1,7 @@
 var expect = require('expect.js');
 var sinon = require('sinon');
 
-var env = require('../../importer.js');
+var env = require('../../env.js');
 
 var extend = env.Engine.Util.extend;
 var Host = env.Engine.Object;
